pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-2-whisper"
version = "0.1.0"
description = "YouTube to Whisper transcription tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
    "yt-dlp[default]>=2025.12.8",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"

[project.scripts]
youtube-2-whisper = "youtube_2_whisper.main:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
